+++ cmd/osctl/src/main.rs
@@ -1,6 +1,6 @@
 use clap::{Parser, Subcommand};
-use matic_api::node::node_service_client::NodeServiceClient;
-use matic_api::node::{
+use keel_api::node::node_service_client::NodeServiceClient;
+use keel_api::node::{
     GetHealthRequest, GetRollbackHistoryRequest, GetStatusRequest, InstallUpdateRequest,
     RebootRequest, TriggerRollbackRequest,
 };
@@ -8,7 +8,7 @@ use tokio_stream::StreamExt;
 
 #[derive(Parser)]
 #[command(name = "osctl")]
-#[command(about = "MaticOS CLI Client", long_about = None)]
+#[command(about = "KeelOS CLI Client", long_about = None)]
 struct Cli {
     #[arg(long, default_value = "http://[::1]:50051")]
     endpoint: String,